refactor(specgen): split generator.go into collect/index/writer; split tests to match; fix .yamllint path
Split generator.go into focused files:
  collect.go  — read steplib source into []parsedStep
  index.go    — build spec/ index files and write per-step files
  writer.go   — file-emitting writer with byte/file accounting

Split generator_test.go to match: collect_test.go and index_test.go.
Test names follow Test<File>_<scenario> convention.

Also fix .yamllint ignore path (specv2/testdata → steplibrary/specgen/testdata)
and add exhaustruct zero values to writer literal.
