+++ steplibrary/specgen/collect_test.go
@@ -1,0 +1,152 @@
+package specgen
+
+import (
+	"os"
+	"path/filepath"
+	"testing"
+	"testing/fstest"
+
+	"github.com/bitrise-io/stepman/models"
+	"github.com/bitrise-io/stepman/steplibrary/spec"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestCollect_step_info_and_asset_copy(t *testing.T) {
+	out := runGenerateFromSteplibClone(t)
+
+	var info spec.StepInfo
+	readJSON(t, filepath.Join(out, "steps/hello-step/step-info.json"), &info)
+
+	assert.Equal(t, "bitrise", info.Maintainer, "Maintainer")
+	assert.Nil(t, info.Deprecation, "Deprecation")
+	assert.Equal(t, map[string]string{"icon.svg": "assets/icon.svg"}, info.AssetURLs, "AssetURLs")
+
+	// Asset file copied.
+	_, gotErr := os.Stat(filepath.Join(out, "steps/hello-step/assets/icon.svg"))
+	assert.NoError(t, gotErr, "asset file copied")
+}
+
+func TestCollect_asset_permissions_preserved(t *testing.T) {
+	out := runGenerateFromSteplibClone(t)
+
+	src := "testdata/input/steps/hello-step/assets/icon.svg"
+	dst := filepath.Join(out, "steps/hello-step/assets/icon.svg")
+
+	srcInfo, err := os.Stat(src)
+	require.NoError(t, err, "stat source asset")
+	dstInfo, err := os.Stat(dst)
+	require.NoError(t, err, "stat copied asset")
+
+	assert.Equal(t, srcInfo.Mode(), dstInfo.Mode(), "copied asset preserves source file permissions")
+}
+
+func TestCollect_deprecated_step(t *testing.T) {
+	out := runGenerateFromSteplibClone(t)
+
+	var info spec.StepInfo
+	readJSON(t, filepath.Join(out, "steps/deprecated-step/step-info.json"), &info)
+
+	assert.Equal(t, "community", info.Maintainer, "Maintainer")
+	require.NotNil(t, info.Deprecation, "Deprecation")
+	assert.Equal(t, "2026-12-31", info.Deprecation.RemovalDate, "RemovalDate")
+	assert.Contains(t, info.Deprecation.Notes, "Replaced by hello-step", "Notes")
+	assert.Empty(t, info.AssetURLs, "no assets dir → no asset_urls")
+}
+
+func TestCollect_no_info_step_skips_step_info_file(t *testing.T) {
+	inputFS := fstest.MapFS{
+		"steplib.yml":                       {Data: []byte("format_version: '0.9.0'\n")},
+		"steps/no-info-step/1.0.0/step.yml": {Data: []byte("title: No Info\n")},
+	}
+	out := t.TempDir()
+	_, gotErr := GenerateFromSteplibClone(inputFS, out, Options{GeneratedAt: fixedTime}, testLogger{t})
+	require.NoError(t, gotErr, "GenerateFromSteplibClone")
+
+	// step-info.json must NOT exist: no step-info.yml and no assets.
+	_, statErr := os.Stat(filepath.Join(out, "steps/no-info-step/step-info.json"))
+	assert.True(t, os.IsNotExist(statErr), "step-info.json should not exist; got err=%v", statErr)
+
+	// step.json must still be written.
+	_, statErr = os.Stat(filepath.Join(out, "steps/no-info-step/1.0.0/step.json"))
+	assert.NoError(t, statErr, "step.json written")
+}
+
+func TestCollect_step_info_written_for_assets_only_step(t *testing.T) {
+	// step-info.json should be written when a step has assets but no step-info.yml.
+	inputFS := fstest.MapFS{
+		"steplib.yml":                           {Data: []byte("format_version: '0.9.0'\nsteplib_source: 'https://example.com'\n")},
+		"steps/asset-only-step/1.0.0/step.yml":  {Data: []byte("title: Asset Only\n")},
+		"steps/asset-only-step/assets/icon.svg": {Data: []byte("<svg/>"), Mode: 0o644},
+	}
+	out := t.TempDir()
+	_, gotErr := GenerateFromSteplibClone(inputFS, out, Options{GeneratedAt: fixedTime}, testLogger{t})
+	require.NoError(t, gotErr, "GenerateFromSteplibClone")
+
+	var info spec.StepInfo
+	readJSON(t, filepath.Join(out, "steps/asset-only-step/step-info.json"), &info)
+	assert.Equal(t, map[string]string{"icon.svg": "assets/icon.svg"}, info.AssetURLs, "AssetURLs")
+	assert.Empty(t, info.Maintainer, "Maintainer")
+	assert.Nil(t, info.Deprecation, "Deprecation")
+}
+
+func TestCollect_invalid_version_dir_skipped(t *testing.T) {
+	inputFS := fstest.MapFS{
+		"steplib.yml":                            {Data: []byte("format_version: '0.9.0'\n")},
+		"steps/my-step/1.0.0/step.yml":           {Data: []byte("title: My Step\n")},
+		"steps/my-step/not-a-semver/step.yml":    {Data: []byte("title: Should be skipped\n")},
+		"steps/my-step/also-not-semver/step.yml": {Data: []byte("title: Also skipped\n")},
+	}
+	out := t.TempDir()
+	stats, gotErr := GenerateFromSteplibClone(inputFS, out, Options{GeneratedAt: fixedTime}, testLogger{t})
+	require.NoError(t, gotErr, "GenerateFromSteplibClone")
+
+	assert.Equal(t, 1, stats.StepCount, "StepCount")
+	assert.Equal(t, 1, stats.VersionCount, "VersionCount")
+
+	_, statErr := os.Stat(filepath.Join(out, "steps/my-step/1.0.0/step.json"))
+	assert.NoError(t, statErr, "valid version written")
+	_, statErr = os.Stat(filepath.Join(out, "steps/my-step/not-a-semver/step.json"))
+	assert.True(t, os.IsNotExist(statErr), "non-semver version dir skipped")
+}
+
+func TestCollect_multi_platform_executables(t *testing.T) {
+	out := runGenerateFromSteplibClone(t)
+
+	var step models.StepModel
+	readJSON(t, filepath.Join(out, "steps/multi-platform-step/3.2.1/step.json"), &step)
+
+	require.NotNil(t, step.Executables, "Executables")
+	require.Len(t, *step.Executables, 2, "Executables len")
+
+	darwinArm := (*step.Executables)["darwin-arm64"]
+	assert.Equal(t,
+		"steps/multi-platform-step/3.2.1/bin/multi-platform-step-darwin-arm64",
+		darwinArm.StorageURI,
+		"darwin-arm64 StorageURI",
+	)
+	assert.Equal(t,
+		"sha256-1111aaaa1111aaaa1111aaaa1111aaaa1111aaaa1111aaaa1111aaaa1111aaaa",
+		darwinArm.Hash,
+		"darwin-arm64 Hash",
+	)
+
+	linuxAmd := (*step.Executables)["linux-amd64"]
+	assert.Equal(t,
+		"steps/multi-platform-step/3.2.1/bin/multi-platform-step-linux-amd64",
+		linuxAmd.StorageURI,
+		"linux-amd64 StorageURI",
+	)
+}
+
+func TestCollect_bash_step_has_no_executables(t *testing.T) {
+	out := runGenerateFromSteplibClone(t)
+
+	var step models.StepModel
+	readJSON(t, filepath.Join(out, "steps/bash-step/1.0.0/step.json"), &step)
+
+	assert.Nil(t, step.Executables, "Executables")
+	require.NotNil(t, step.Toolkit, "Toolkit")
+	require.NotNil(t, step.Toolkit.Bash, "Toolkit.Bash")
+	assert.Equal(t, "step.sh", step.Toolkit.Bash.EntryFile, "Toolkit.Bash.EntryFile")
+}
